+++ sync-for-commerce/CodatSyncCommerce/CodatSyncCommerce.cs
@@ -17,6 +17,8 @@ namespace Codat.Sync.Commerce
     using System.Threading.Tasks;
     using System;
 
+
+
     /// <summary>
     /// Sync for Commerce: The API for Sync for Commerce. <br/>
     /// 
@@ -60,6 +62,22 @@ namespace Codat.Sync.Commerce
     
     public class SDKConfig
     {
+        public static string[] ServerList = new string[]
+        {
+            "https://api.codat.io",
+        };
+        /// Contains the list of servers available to the SDK
+        public string serverUrl = "";
+        public int serverIndex = 0;
+
+        public string GetTemplatedServerDetails()
+        {
+            if (!String.IsNullOrEmpty(this.serverUrl))
+            {
+                return Utilities.TemplateUrl(Utilities.RemoveSuffix(this.serverUrl, "/"), new Dictionary<string, string>());
+            }
+            return Utilities.TemplateUrl(SDKConfig.ServerList[this.serverIndex], new Dictionary<string, string>());
+        }
     }
 
     /// <summary>
@@ -77,16 +95,12 @@ namespace Codat.Sync.Commerce
     public class CodatSyncCommerce: ICodatSyncCommerce
     {
         public SDKConfig Config { get; private set; }
-        public static List<string> ServerList = new List<string>()
-        {
-            "https://api.codat.io",
-        };
 
         private const string _language = "csharp";
-        private const string _sdkVersion = "3.0.0";
-        private const string _sdkGenVersion = "2.188.1";
+        private const string _sdkVersion = "3.0.1";
+        private const string _sdkGenVersion = "2.194.1";
         private const string _openapiDocVersion = "1.1";
-        private const string _userAgent = "speakeasy-sdk/csharp 3.0.0 2.188.1 1.1 Codat.Sync.Commerce";
+        private const string _userAgent = "speakeasy-sdk/csharp 3.0.1 2.194.1 1.1 Codat.Sync.Commerce";
         private string _serverUrl = "";
         private ISpeakeasyHttpClient _defaultClient;
         private ISpeakeasyHttpClient _securityClient;
@@ -96,9 +110,14 @@ namespace Codat.Sync.Commerce
         public ISync Sync { get; private set; }
         public IIntegrations Integrations { get; private set; }
 
-        public CodatSyncCommerce(Security? security = null, string? serverUrl = null, ISpeakeasyHttpClient? client = null)
+        public CodatSyncCommerce(Security? security = null, int? serverIndex = null, string? serverUrl = null, Dictionary<string, string>? urlParams = null, ISpeakeasyHttpClient? client = null)
         {
-            _serverUrl = serverUrl ?? CodatSyncCommerce.ServerList[0];
+            if (serverUrl != null) {
+                if (urlParams != null) {
+                    serverUrl = Utilities.TemplateUrl(serverUrl, urlParams);
+                }
+                _serverUrl = serverUrl;
+            }
 
             _defaultClient = new SpeakeasyHttpClient(client);
             _securityClient = _defaultClient;
@@ -110,6 +129,7 @@ namespace Codat.Sync.Commerce
             
             Config = new SDKConfig()
             {
+                serverUrl = _serverUrl
             };
 
             SyncFlowSettings = new SyncFlowSettings(_defaultClient, _securityClient, _serverUrl, Config);