+++ sync-for-commerce/CodatSyncCommerce/Models/Operations/ListSyncsResponse.cs
@@ -36,7 +36,7 @@ namespace Codat.Sync.Commerce.Models.Operations
         /// <summary>
         /// Raw HTTP response; suitable for custom response parsing
         /// </summary>
-        public HttpResponseMessage? RawResponse { get; set; }
+        public HttpResponseMessage RawResponse { get; set; } = default!;
 
         /// <summary>
         /// Success